refactor: addauthor server.js move to authors/routes.js working
